src/cli/user.command.spec.ts
/*
 * SPDX-FileCopyrightText: 2021 The HedgeDoc developers (see AUTHORS file)
 *
 * SPDX-License-Identifier: AGPL-3.0-only
 */

import { Test } from '@nestjs/testing';
import { CommandModule, CommandModuleTest } from 'nestjs-command';
import { UserCommand } from './user.command';
import { UsersService } from '../users/users.service';
import { getRepositoryToken } from '@nestjs/typeorm';
import { User } from '../users/user.entity';
import { Repository } from 'typeorm';

describe('UserCommand', () => {
  let commandModule: CommandModuleTest;

  beforeEach(async () => {
    const moduleFixture = await Test.createTestingModule({
      providers: [
        UsersService,
        {
          provide: getRepositoryToken(User),
          useClass: Repository,
        },
      ],
      imports: [CommandModule, UserCommand],
    }).compile();

    const app = moduleFixture.createNestApplication();
    await app.init();
    commandModule = new CommandModuleTest(app.select(CommandModule));
  });

  it('config show works', async () => {
    const consoleSpy = jest.spyOn(console, 'log');

    const command = 'config show';
    const args = {};
    const exitCode = 0;

    await commandModule.execute(command, args, exitCode);
    expect(true).toBe(true);
    /*
    expect(consoleSpy).toHaveBeenCalledWith('AppConfig: ', appConfigMock());
    expect(consoleSpy).toHaveBeenCalledWith('AuthConfig: ', authConfig());
    expect(consoleSpy).toHaveBeenCalledWith('CspConfig: ', cspConfig());
    expect(consoleSpy).toHaveBeenCalledWith(
      'DatabaseConfig: ',
      databaseConfig(),
    );
    expect(consoleSpy).toHaveBeenCalledWith('HstsConfig: ', hstsConfig());
    expect(consoleSpy).toHaveBeenCalledWith('MediaConfig: ', mediaConfigMock());
    */
  });
});
